requirements.txt
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126
pytorch-lightning
wandb
tqdm
numpy
torchsummary
torchinfo
datasets